percen

+++ src/main/java/com/lawencon/psikotest/dao/PackageDetailDao.java
@@ -140,14 +140,15 @@ public class PackageDetailDao extends EntityDao {
 							+ " limit 5");
 			List<BigInteger> point = queryPoint.getResultList();
 			
-			Integer totalQuestion = countQuestion(p.getPackageId()).intValue();
-			Integer percentage;
+			Double totalQuestion = countQuestion(p.getPackageId()).doubleValue();
+			Double percentage;
+			
 			for(int i=0; i<question.size(); i++) {
 				POJOStats cs = new POJOStats();
 				cs.setPackageName(pack.get(i));
 				cs.setQuestion(question.get(i));
 				cs.setCorrect(point.get(i).intValue());
-				percentage = (point.get(i).intValue()/totalQuestion) * 100;
+				percentage = (point.get(i).doubleValue()/totalQuestion) * 100;
 				cs.setPercentage(percentage);
 				stats.add(cs);
 			}
